added games and routes

+++ routes/api-routes.js
@@ -18,45 +18,49 @@ module.exports = function (app,io){
       })
 
     app.post('/upload',function(req,res){
-        //Initialize an empty form
-        var form = new formidable.IncomingForm()
-        form.maxFileSize = Math.pow(1024,3)
-        form.parse(req,function(err,fields,files){
-            console.log(fields)
-            if(err) throw err;
-            //fields are any values that are not files that were sent with the form whose keys are the name of the input
-            //Grab the path of the file that was just uploaded "filetoupload" is the name of the input on the frontend
-            var oldpath = files.filetoupload.path;
-            console.log(files.thumbnail.path)
-            //Create a newpath to store the file at
-            var newpath = __dirname + "/" + files.filetoupload.name
+        if(req.user){
+            console.log(req.user)
+            //Initialize an empty form
+            var form = new formidable.IncomingForm()
+            form.maxFileSize = Math.pow(1024,3)
+            form.parse(req,function(err,fields,files){
+                console.log(fields)
+                if(err) throw err;
+                //fields are any values that are not files that were sent with the form whose keys are the name of the input
+                //Grab the path of the file that was just uploaded "filetoupload" is the name of the input on the frontend
+                var oldpath = files.filetoupload.path;
+                console.log(files.thumbnail.path)
+                //Create a newpath to store the file at
+                var newpath = __dirname + "/" + files.filetoupload.name
 
-            fs.rename(oldpath, newpath, function (err) {
-                console.log('renaming filepaths')
-                if (err) throw err;
+                fs.rename(oldpath, newpath, function (err) {
+                    console.log('renaming filepaths')
+                    if (err) throw err;
 
-                //Create a directory if it doesn't already exist
-                var dir = "./client/public/user"
-                if (!fs.existsSync(dir)){
-                    console.log('path does not exist \n creating newpath')
-                    fs.mkdirSync(dir);
-                }
+                    //Create a directory if it doesn't already exist
+                    var dir = "./client/public/" + req.user.id + '/' + fields.title
+                    if (!fs.existsSync(dir)){
+                        console.log('path does not exist \n creating newpath')
+                        fs.mkdirSync(dir);
+                    }
 
-                //Unzip the file to target directory
-                var target = path.join(__dirname,'../client/public/user/')
-                extract(newpath,{dir:target},function(err){
-                    console.log('extracting to ', target)
-                    if(err) throw err;
-                    fs.unlink(newpath, (err) => {
-                        if (err) throw err;
-                        console.log('deleting' + newpath );
-                        //Send the URL to the games new path
-                        res.send('/game');
-                    });
-                })
-            });
-        })
-    
+                    //Unzip the file to target directory
+                    var target = path.join(__dirname,'../client/public/' + req.user.id + '/' + fields.title)
+                    extract(newpath,{dir:target},function(err){
+                        console.log('extracting to ', target)
+                        if(err) throw err;
+                        fs.unlink(newpath, (err) => {
+                            if (err) throw err;
+                            console.log('deleting' + newpath );
+                            //Send the URL to the games new path
+                            res.send('/game');
+                        });
+                    })
+                });
+            })
+
+        }
+        
     })
 
     app.post("/api/login",passport.authenticate("local"),  function(req, res) {
